Separate algo components
