+++ api/ai/new/geg_algorithm/geg_algorithm.py
@@ -1,7 +1,8 @@
 import math
 import random
-from typing import List, Dict, Tuple, Set
+from typing import List, Dict, Tuple
 
+from api.ai.new.geg_algorithm.envy_graph import EnvyGraph
 from api.ai.new.interfaces.algorithm import Algorithm
 from api.ai.new.interfaces.algorithm_options import GeneralizedEnvyGraphAlgorithmOptions
 from api.ai.new.interfaces.team_generation_options import TeamGenerationOptions
@@ -10,7 +11,6 @@ from api.models.enums import (
     ScenarioAttribute,
     Gender,
     Race,
-    RequirementOperator,
 )
 from api.models.project import Project, ProjectRequirement
 from api.models.student import Student
@@ -21,159 +21,6 @@ from benchmarking.data.simulated_data.mock_student_provider import (
     MockStudentProvider,
 )
 from benchmarking.evaluations.graphing.graph_metadata import GraphData
-
-
-def get_additive_utilities(
-    students: List[Student], projects: List[Project]
-) -> Dict[Tuple[int, int], int]:
-    utilities: Dict[Tuple[int, int], int] = {}
-
-    for project in projects:
-        project_requirements = set(project.requirements)
-        for student in students:
-            student_utilities = 0
-            student_attributes = student.attributes
-            for attribute_id, attribute_values in student_attributes.items():
-                if attribute_id in project_requirements:
-                    student_utilities += sum(
-                        [
-                            1
-                            for attribute_value in attribute_values
-                            if attribute_value > 0
-                        ]
-                    )
-                else:
-                    student_utilities -= 1
-
-            utilities[(project.id, student.id)] = student_utilities
-
-    return utilities
-
-
-class EnvyGraph:
-    def __init__(self, students: List[Student]) -> None:
-        """
-        Initalize an envy graph
-
-        This run in O(N)
-        """
-        vertices = [student.id for student in students]
-        self.graph: Dict[int, Set[int]] = {vertex: set([]) for vertex in vertices}
-        self.additive_utilities: Dict[int, int] = {vertex: 0 for vertex in vertices}
-
-    def add_edge(self, envy_project: int, other_project: int) -> None:
-        """
-        This run in O(1)
-        """
-        self.graph[envy_project].add(other_project)
-
-    def remove_edge(self, no_loger_envy_project: int, other_project: int) -> None:
-        """
-        This run in O(1)
-        """
-        if other_project in self.graph[no_loger_envy_project]:
-            self.graph[no_loger_envy_project].remove(other_project)
-
-    def get_envy_targets(self, envy_student: int) -> Set[int]:
-        """
-        This run in O(1)
-        """
-
-        return self.graph[envy_student]
-
-    def is_sink(self, student: int) -> bool:
-        """
-        A student is sink if there is no outgoing edge
-
-        This run in O(1)
-        """
-        return len(self.graph[student]) == 0
-
-    def is_source(self, student: int) -> bool:
-        """
-        A student is source if there is no incoming edge
-
-        This run in O(N)
-        """
-        return all([student not in targets for targets in self.graph.values()])
-
-    def update_envy_graph(self, project_id: int, new_utility_value: int) -> None:
-        """
-        Construct a new envy graph induced by N
-
-        This run in O(N)
-        """
-
-        # Since this is additive, add new utility value to the existed value
-        self.additive_utilities[project_id] += new_utility_value
-
-        updated_utility = self.additive_utilities[project_id]
-        for other_project_id in self.additive_utilities.keys():
-            if project_id == other_project_id:
-                continue
-
-            # In case utility increases, no loger envy with some and some will start envy
-            if updated_utility > self.additive_utilities[other_project_id]:
-                self.remove_edge(project_id, other_project_id)
-                self.add_edge(other_project_id, project_id)
-
-            # In case utility decreases, will start envy with more and some will stop envying
-            if updated_utility < self.additive_utilities[other_project_id]:
-                self.add_edge(project_id, other_project_id)
-                self.remove_edge(other_project_id, project_id)
-
-            # In case utility is equal to other, noone will envy
-            if updated_utility == self.additive_utilities[other_project_id]:
-                self.remove_edge(project_id, other_project_id)
-                self.remove_edge(other_project_id, project_id)
-
-    def get_all_cycles(self) -> List[List[int]]:
-        visited = set([])
-        cycles = []
-
-        for start_node in self.graph:
-            self._dfs(start_node, visited, [])
-
-        return cycles
-
-    def _dfs(self, node: int, visited: Set[int], path: List[int]) -> List[int]:
-        if node in path:
-            return path[path.index(node) :]
-
-        if node not in visited:
-            visited.add(node)
-            path.append(node)
-            for neighbor in self.get_envy_targets(node):
-                self._dfs(neighbor, visited, path)
-            path.pop()
-            visited.remove(node)
-
-    def draw(self) -> None:
-        import networkx as nx
-        import matplotlib.pyplot as plt
-
-        G = nx.DiGraph()
-        G.add_nodes_from(self.graph.keys())
-        for source, targets in self.graph.items():
-            for target in targets:
-                G.add_edge(source, target)
-
-        nx.draw(G, with_labels=True)
-        plt.show()
-
-
-def requirement_met_by_student(
-    requirement: ProjectRequirement, student: Student
-) -> bool:
-    is_met = False
-    for value in student.attributes.get(requirement.attribute):
-        if requirement.operator == RequirementOperator.LESS_THAN:
-            is_met |= value < requirement.value
-        elif requirement.operator == RequirementOperator.MORE_THAN:
-            is_met |= value > requirement.value
-        else:  # default case is 'exactly'
-            is_met |= value == requirement.value
-    return is_met
 
 
 class GEGAlgorithm(Algorithm):
@@ -187,19 +34,18 @@ class GEGAlgorithm(Algorithm):
     algorithm_run_time: float
 
     def __init__(
-        self,
-        algorithm_options: GeneralizedEnvyGraphAlgorithmOptions,
-        team_generation_options: TeamGenerationOptions,
-        projects: List[Project],
+            self,
+            algorithm_options: GeneralizedEnvyGraphAlgorithmOptions,
+            team_generation_options: TeamGenerationOptions,
     ):
         super().__init__(algorithm_options, team_generation_options)
 
         self.allocation: Dict[int, List[int]] = {}
         self.utilities: Dict[Tuple[int, int], int] = {}
-        self.projects: List[Project] = projects
+        self.projects: List[Project] = algorithm_options.projects
 
     def _randomize_utilities(
-        self, students: List[Student]
+            self, students: List[Student]
     ) -> Dict[Tuple[int, int], int]:
         utilities: Dict[Tuple[int, int], int] = {}
 
@@ -356,7 +202,7 @@ if __name__ == "__main__":
             projects.append(Project(proj_id, requirements=proj_requirements))
 
         geg_algorithm = GEGAlgorithm(
-            GeneralizedEnvyGraphAlgorithmOptions(), None, projects
+            GeneralizedEnvyGraphAlgorithmOptions(projects=projects), None
         )
         team_set = geg_algorithm.generate(student_provider.get())
