+++ api/ai/new/interfaces/algorithm_options.py
@@ -90,5 +90,11 @@ class MultipleRoundRobinAlgorithmOptions(AlgorithmOptions):
 
 @dataclass
 class GeneralizedEnvyGraphAlgorithmOptions(AlgorithmOptions):
+    projects: List[Project]
+
     def validate(self):
         super().validate()
+
+        Schema([Project]).validate(self.projects)
+        if len(self.projects) == 0:
+            raise SchemaError("Project list cannot be empty")
